move display_utils to qsc_data_structures
