+++ source/compiler/qsc_data_structures/Cargo.toml
@@ -9,6 +9,7 @@ license.workspace = true
 repository.workspace = true
 
 [dependencies]
+indenter = { workspace = true }
 index_map = { path = "../../index_map" }
 miette = { workspace = true }
 serde = { workspace = true }
